used propTypes, created components folder
