+++ src/components/InputWithLabel.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import './TodoListItem.module.css';
 
 function InputWithLabel(props) {
@@ -14,6 +15,10 @@ function InputWithLabel(props) {
             <input ref={inputRef} id='todoTitle' value={props.value} type='text' name='title' onChange={props.onChange} />
         </>
     );
+}
+
+InputWithLabel.propTypes = {
+    onChange: PropTypes.func,
 }
 
 export default InputWithLabel
